Rescheduling files after the new servents have joined the party.

+++ src/servent/handler/NewNodeHandler.java
@@ -7,15 +7,13 @@ import servent.message.*;
 import servent.message.util.MessageUtil;
 
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 import java.util.Map.Entry;
 
 public class NewNodeHandler implements MessageHandler {
 
     private Message clientMessage;
+    private List<Integer> processedHashValues = Collections.synchronizedList(new ArrayList<>());
 
     public NewNodeHandler(Message clientMessage) {
         this.clientMessage = clientMessage;
@@ -80,25 +78,41 @@ public class NewNodeHandler implements MessageHandler {
                     }
 
                 }
-                for (Integer key : hisValues.keySet()) { //remove his values from my map
+                for (Integer key : hisValues.keySet()) { // remove his values from my map
                     myValues.remove(key);
                     SillyFile sillyFile = hisValues.get(key);
-                    int hash = key;
+
                     if (sillyFile.isDirectory()) {
-                        // TODO: Send file by file
-//                        sendFiles(sillyFile, AppConfig.myServentInfo, newNodeInfo, hash, newNodePort);
+                        for (Map.Entry<Integer, SillyFile> entry : sillyFile.getSillyFiles().entrySet()) {
+                            SillyFile subFile = entry.getValue();
+                            int hash = entry.getKey();
+                            if (!processedHashValues.contains(hash)) {
+                                AddMessage addMessage = new AddMessage(AppConfig.myServentInfo.getListenerPort(), newNodePort, String.valueOf(hash), subFile, false);
+                                MessageUtil.sendMessage(addMessage);
+                                if (hisValues.containsKey(hash)) {
+                                    // Note: Let files exist separately if their hash belongs to the new node
+                                    File file = new File(AppConfig.myServentInfo.getStorage() + subFile.getFilePath());
+                                    file.delete();
+                                }
+                                processedHashValues.add(hash);
+                            }
+                        }
                     } else {
-                        AddMessage addMessage = new AddMessage(AppConfig.myServentInfo.getListenerPort(), newNodePort, String.valueOf(hash), sillyFile, false);
-                        MessageUtil.sendMessage(addMessage);
-                        File file = new File(AppConfig.myServentInfo.getStorage() + sillyFile.getFilePath());
-                        file.delete();
+                        int hash = key;
+                        if (!processedHashValues.contains(hash)) {
+                            AddMessage addMessage = new AddMessage(AppConfig.myServentInfo.getListenerPort(), newNodePort, String.valueOf(hash), sillyFile, false);
+                            MessageUtil.sendMessage(addMessage);
+                            File file = new File(AppConfig.myServentInfo.getStorage() + sillyFile.getFilePath());
+                            file.delete();
+                            processedHashValues.add(hash);
+                        }
                     }
                 }
                 AppConfig.chordState.setValueMap(myValues);
 
                 WelcomeMessage wm = new WelcomeMessage(AppConfig.myServentInfo.getListenerPort(), newNodePort, hisValues);
                 MessageUtil.sendMessage(wm);
-            } else { //if he is not my predecessor, let someone else take care of it
+            } else { // if he is not my predecessor, let someone else take care of it
                 ServentInfo nextNode = AppConfig.chordState.getNextNodeForKey(newNodeInfo.getChordId());
                 NewNodeMessage nnm = new NewNodeMessage(newNodePort, nextNode.getListenerPort());
                 MessageUtil.sendMessage(nnm);
@@ -106,41 +120,6 @@ public class NewNodeHandler implements MessageHandler {
 
         } else {
             AppConfig.timestampedErrorPrint("NEW_NODE handler got something that is not new node message.");
-        }
-    }
-
-    private void sendFiles(SillyFile sillyFile, ServentInfo myServentInfo, ServentInfo newNodeInfo, int hash, int newNodePort) {
-        // TODO: won't work
-        try {
-            if (sillyFile.isDirectory()) {
-                File dir = new File(myServentInfo.getStorage() + sillyFile.getDirectoryPath());
-                for (File file : dir.listFiles()) {
-                    String path = sillyFile.getDirectoryPath() + "\\" + file.getName();
-                    if (!file.isDirectory()) {
-                        SillyFile sFile = new SillyFile(Files.readAllBytes(file.toPath()), path);
-                        sendFiles(sFile, myServentInfo, newNodeInfo, hash, newNodePort);
-                    } else {
-                        File storageFile = new File(myServentInfo.getStorage() + "\\" + path);
-                        File dirForFile = new File(storageFile.getPath());
-                        if (!dirForFile.exists()) dirForFile.mkdirs();
-                        sendFiles(new SillyFile(path), myServentInfo, newNodeInfo, hash, newNodePort);
-                    }
-                }
-            } else {
-                byte[] fileContent = sillyFile.getFileContent();
-                String storageFileName = AppConfig.myServentInfo.getStorage() + sillyFile.getFilePath();
-                File storageFile = new File(storageFileName);
-                File dir = new File(storageFile.getParent());
-                if (!dir.exists()) dir.mkdirs();
-                Files.write(storageFile.toPath(), fileContent);
-
-                AddMessage addMessage = new AddMessage(AppConfig.myServentInfo.getListenerPort(), newNodePort, String.valueOf(hash), sillyFile, false);
-                MessageUtil.sendMessage(addMessage);
-                File file = new File(AppConfig.myServentInfo.getStorage() + sillyFile.getFilePath());
-                file.delete();
-            }
-        } catch (IOException e) {
-            AppConfig.timestampedErrorPrint("Couldn't add file [" + sillyFile.getFilePath() + "] to storage.");
         }
     }
 }
